// admin_backend/src/types/mongoose-types.ts
export type ShortendUrl = {
  link_title: string;
  link_description?: string;
  original_url: string;
  link_enabeld: boolean;
  shortend_url_cuid: string;
  creator_id: User["_id"] | User;
  link_cloaking: boolean;
  sharing_preview: {
    title?: string;
    description?: string;
    image?: string;
  };
  tags: string[];
  protected: {
    enabeld: boolean;
    password?: string;
  };
  link_expiry: {
    enabeld: boolean;
    link_expires_on?: string;
    expiry_redirect_url?: string;
  };
  link_targetting: {
    enabeld: boolean;
    location: {
      country?: string;
      redirect_url?: string;
    };
    device: {
      ios?: string;
      android?: string;
      windows?: string;
      linux?: string;
      mac?: string;
    };
    rotate: string[];
  };
  stats?: Stat["_id"] | Stat;
  _id: mongoose.Types.ObjectId;
  createdAt?: Date;
  updatedAt?: Date;
};
// ...
export type SubscriptionPurchase_log = {
  date_of_purchase: string;
  expired_on: string;
  product_id: string;
  product_name: string;
  price_id: string;
  amount: number;
  payment_method_brand: string;
  card_last4: string;
  _id: mongoose.Types.ObjectId;
};

/**
 * Lean version of SubscriptionDocument
 *
 * This has all Mongoose getters & functions removed. This type will be returned from `SubscriptionDocument.toObject()`. To avoid conflicts with model names, use the type alias `SubscriptionObject`.
 * ```
 * const subscriptionObject = subscription.toObject();
 * ```
 */
export type Subscription = {
  customer_stripe_id: string;
  stripe_subscription_id: string;
  user_id: User["_id"] | User;
  product_name: "trial" | "personal" | "team" | "enterprise";
  usuage: {
    link_generated: number;
    custom_domains: number;
    landing_pages: number;
    workspaces: number;
    teams: number;
    last_interval_date: string;
  };
  product_id: string;
  plan_name: "trial" | "monthly" | "annual";
  price_id: string;
  price: number;
  currency: "ruppee" | "dollar";
  interval_value: number;
  interval_decimal: "day" | "month" | "year";
  status: "OK" | "PAYMENT METHOD ERROR" | "PLAN ENDED";
  error?: string;
  purchase_log: SubscriptionPurchase_log[];
  valid_till: string;
  _id: mongoose.Types.ObjectId;
  createdAt?: Date;
  updatedAt?: Date;
};
// ...
export type UserPayment_method = {
  pm_type: string;
  brand: string;
  last_4_card_digits: string;
  expiry_month: number;
  expiry_year: number;
  _id: mongoose.Types.ObjectId;
};
// ...
export type UserBilling_address = {
  city: string;
  line1: string;
  line2: string | null;
  postal_code: string;
  state: string;
  _id: mongoose.Types.ObjectId;
};

/**
 * Lean version of UserDocument
 *
 * This has all Mongoose getters & functions removed. This type will be returned from `UserDocument.toObject()`. To avoid conflicts with model names, use the type alias `UserObject`.
 * ```
 * const userObject = user.toObject();
 * ```
 */
export type User = {
  name: string;
  email: string;
  profile_img: string;
  password: string;
  subscription_id?: Subscription["_id"] | Subscription;
  payment_method: UserPayment_method[];
  generated_links: (ShortendUrl["_id"] | ShortendUrl)[];
  favorites: (ShortendUrl["_id"] | ShortendUrl)[];
  address?: UserBilling_address;
  billing_address?: UserBilling_address;
  googleOAuthId: string;
  githubOAuthId: string;
  customerStripeId: string;
  _id: mongoose.Types.ObjectId;
  createdAt?: Date;
  updatedAt?: Date;
};